type cast warning
git-svn-id: tumanako.svn.sourceforge.net::svn/tumanako/trunk@416 c3429617-7ab9-49b5-8233-5f90b2ce515f

+++ src/c/monitor.c
@@ -684,7 +684,7 @@ void initCellIDArray() {
 
 void findCells() {
 	struct status_t status;
-	for (int i = 0; i < 255; i++) {
+	for (unsigned short i = 0; i < 255; i++) {
 		status.cellId = i;
 		status.version = 0;
 		unsigned char found = 0;
